agent-image-discussion/requirements.txt
autogen-core==0.5.6
autogen-ext[openai,azure]==0.5.6
autogen-agentchat==0.5.6
python-dotenv==1.1.0